CR#11110 adding eager mode to profiler to support Azure webapps
These applications will be killed instead of properly shut down, so the
profiler cannot write the trace infos in it's Shutdown method. Instead,
eager mode enables writing of the trace file as soon as events are
received from the CLR.

The resulting trace files may have a different order of entries (e.g.
Assembly= statements interleaved with Inlined= or Jitted=).